pyproject.toml
[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[project]
name = "kabukit"
version = "0.1.2"
description = "A Python toolkit for Japanese financial market data, supporting J-Quants and EDINET APIs."
readme = "README.md"
authors = [{ name = "kabukit", email = "kabukitlab@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
  "altair>=5.5.0",
  "async-typer>=0.1.10",
  "holidays>=0.81",
  "httpx>=0.28.1",
  "marimo[lsp]>=0.16.1",
  "platformdirs>=4.4.0",
  "polars>=1.33.1",
  "python-dotenv>=1.1.1",
  "typer>=0.19.2",
  "vegafusion-python-embed>=1.6.9",
  "vegafusion>=2.0.2",
  "vl-convert-python>=1.8.0",
]

[project.scripts]
kabu = "kabukit.cli.app:app"

[dependency-groups]
dev = [
  "basedpyright>=1.31.4",
  "mkdocs-marimo>=0.2.1",
  "mkdocs-material>=9.6.20",
  "mkdocs>=1.6.1",
  "numpy>=2.3.3",            # polars 1.33 type hinting workaround,
  "pytest-asyncio>=1.2.0",
  "pytest-clarity>=1.0.1",
  "pytest-cov>=7.0.0",
  "pytest-randomly>=4.0.1",
  "pytest-xdist>=3.8.0",
  "rich>=14.1.0",
  "tqdm>=4.67.1",
]
docs = ["mkapi>=4.4", "mkdocs-material"]

[tool.pytest.ini_options]
addopts = ["--cov=kabukit", "--cov-report=lcov:lcov.info", "--doctest-modules"]
testpaths = ["src", "tests"]

[tool.coverage.report]
exclude_lines = ["no cov", "raise NotImplementedError", "if TYPE_CHECKING:"]
skip_covered = true

[tool.ruff]
line-length = 88
target-version = "py313"
include = ["src", "tests"]

[tool.ruff.lint]
select = ["ALL"]
unfixable = ["F401"]
ignore = [
  "A002",
  "ANN401",
  "D",
  "FBT001",
  "N802",
  "PGH003",
  "PLC0415",
  "PLR0913",
  "PLR2004",
  "RUF002",
  "S603",
  "SIM102",
  "TRY003",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "FBT", "S101", "S607"]
"schema.py" = ["E501"]
"notebooks/*" = ["F704", "PLE1142"]

[tool.ruff.format]
exclude = ["schema.py"]

[tool.basedpyright]
include = ["src", "tests"]
reportAny = false
reportExplicitAny = false
reportImplicitOverride = false
reportImportCycles = false
reportIncompatibleVariableOverride = false
reportUnusedCallResult = false
reportUnusedImport = false
